Dodaci
- Onemogucen unos ocena leta i kompanije koje su vec ocenjene od strane usera za konkretan let.
- Moguc odabir bonusa prilikom normalne rezervacije.
- Na stranici za obicnu rezervaciju, mogucnost odabira rent a car service.
- Prikaz prosecne ocene (zvezdice) kod leta.

+++ web2-Project/src/app/components/flights-components/fast-flight-reservation/fast-flight-reservation.component.ts
@@ -61,7 +61,7 @@ export class FastFlightReservationComponent implements OnInit {
       this.flight = result as Flight;
       this.ocena = 0;
       for (let index = 0; index < this.flight.ocene.length; index++) {
-        this.ocena += this.flight.ocene[index];
+        this.ocena += this.flight.ocene[index].doubleValue;
       }
       if (this.flight.ocene.length != 0)
         this.ocena = this.ocena / this.flight.ocene.length;
